made application more modular by creating db.js file and exporting functionality where needed to make most request
